Refactor authentication and user-related code
