+++ apps/web/src/scenes/invites/index.tsx
@@ -3,13 +3,12 @@ import Typography from "ui/Typography";
 
 import DashboardWrap from "src/components/DashboardWrap";
 import { useGetInvites } from "src/services/api";
-import { useAppSelector } from "src/services/hooks";
-import { getUser } from "src/services/store";
+import { useAuth } from "src/services/store";
 import InviteListItem from "./components/InviteListItem";
 
 const Invites = () => {
   const navigate = useNavigate();
-  const user = useAppSelector(getUser);
+  const { user } = useAuth();
   const { data: invites, refetch } = useGetInvites();
   if (!user) {
     navigate("/");
