+++ apps/web/src/services/store.ts
@@ -5,6 +5,7 @@ import {
   createSlice,
 } from "@reduxjs/toolkit";
 import auth from "./auth";
+import { useAppDispatch, useAppSelector } from "./hooks";
 
 const initialState: {
   isOpen: boolean;
@@ -47,8 +48,20 @@ export const store = configureStore({
 
 export const getIsLoadingAuth = (state: RootState) => state.auth.isLoadingAuth;
 export const getLoginEmail = (state: RootState) => state.auth.loginEmail;
-export const getUser = (state: RootState) => state.auth.user;
 export const getIsAuthenticated = (state: RootState) => !!state.auth.user;
+
+export const useAuth = () => {
+  const user = useAppSelector((state: RootState) => state.auth.user);
+  const isLoadingAuth = useAppSelector(getIsLoadingAuth);
+  const loginEmail = useAppSelector(getLoginEmail);
+  const dispatch = useAppDispatch();
+  return {
+    user,
+    isLoadingAuth,
+    loginEmail,
+    dispatch,
+  };
+};
 
 export type AppDispatch = typeof store.dispatch;
 export type RootState = ReturnType<typeof store.getState>;
